requirements.txt
# Core dependencies for cryptocurrency pairs trading analysis
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
statsmodels>=0.13.0

# Date manipulation for custom date analysis
python-dateutil>=2.8.0

# Optional dependencies for PDF generation
# pandoc (system installation required)
# texlive-xetex (system installation required)

# Development dependencies (optional)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0